requirements.txt
click==8.0.1
Flask==2.0.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
python-dotenv==0.17.1
Werkzeug==2.0.1
gunicorn==18.0
psycopg2==2.5.1